Correction du back d'après modèle (en cours, a merge en cas de pb)
