+++ birthday-app-back/src/main/java/com/gerikk/birthdayapp/controllers/UserController.java
@@ -3,12 +3,13 @@ package com.gerikk.birthdayapp.controllers;
 import com.gerikk.birthdayapp.exceptions.UserNotFoundException;
 import com.gerikk.birthdayapp.models.Birthday;
 import com.gerikk.birthdayapp.models.User;
-import com.gerikk.birthdayapp.services.BirthdayServiceImpl;
-import com.gerikk.birthdayapp.services.UserServiceImpl;
+import com.gerikk.birthdayapp.services.BirthdayService;
+import com.gerikk.birthdayapp.services.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -17,12 +18,12 @@ import java.util.Set;
 @RequestMapping("/users")
 public class UserController {
 
-    private final UserServiceImpl userService;
+    private final UserService userService;
 
-    private final BirthdayServiceImpl birthdayService;
+    private final BirthdayService birthdayService;
 
 
-    public UserController(UserServiceImpl userService, BirthdayServiceImpl birthdayService) {
+    public UserController(UserService userService, BirthdayService birthdayService) {
         this.userService = userService;
         this.birthdayService = birthdayService;
     }
@@ -41,12 +42,12 @@ public class UserController {
     }
 
     @GetMapping("/{userId}")
-    public User getUserById(@PathVariable("userId") Long id) {
+    public ResponseEntity<User> getUserById(@PathVariable("userId") Long id) {
 
         Optional<User> foundUser = userService.getAllUsers().stream().filter(user -> user.getId().equals(id)).findFirst();
 
         if (foundUser.isPresent()) {
-            return foundUser.get();
+            return ResponseEntity.ok(foundUser.get());
         } else {
             throw new UserNotFoundException();
         }
@@ -62,17 +63,22 @@ public class UserController {
     }
 
     @PostMapping("/{userId}/birthdays")
-    public ResponseEntity<Birthday> createBirthday(@PathVariable("userId") Long id, @RequestBody Birthday newBirthday) {
+    public ResponseEntity<Birthday> createBirthday(
+            @PathVariable("userId") Long id,
+            //@RequestBody Birthday newBirthday
+            @RequestParam("firstname") final String firstname,
+            @RequestParam("lastname") final String lastname,
+            @RequestParam("date") final String date) {
 
-        User user = userService.getUserById(id);
 
-        if (user == null) {
-            throw new UserNotFoundException();
-        } else {
-            newBirthday.setUser(user);
-            user.getBirthdays().add(newBirthday);
-            birthdayService.save(newBirthday);
-            return new ResponseEntity<>(newBirthday, HttpStatus.CREATED);
+        try {
+            //newBirthday.setUser(user);
+            User user = userService.getUserById(id);
+            Birthday newBirthday = new Birthday(user, firstname, lastname, LocalDate.parse(date));
+            //user.getBirthdays().add(newBirthday);
+            return ResponseEntity.ok(birthdayService.save(newBirthday));
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
         }
     }
