+++ birthday-app-back/src/main/java/com/gerikk/birthdayapp/models/Birthday.java
@@ -25,6 +25,9 @@ public class Birthday {
         super();
     }
 
+    public Birthday(User user, String firstname, String lastname, LocalDate date) {
+    }
+
     @Override
     public String toString() {
         return "Birthday{" +
