+++ birthday-app-back/src/main/java/com/gerikk/birthdayapp/services/BirthdayService.java
@@ -1,14 +1,34 @@
 package com.gerikk.birthdayapp.services;
 
 import com.gerikk.birthdayapp.models.Birthday;
+import com.gerikk.birthdayapp.repositories.BirthdayRepository;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface BirthdayService {
+@Service
+public class BirthdayService {
 
-    public List<Birthday> getAllBirthdays();
+    private final BirthdayRepository birthdayRepository;
 
-    public List<Birthday> getBirthdaysByUserId(Long userId);
+    private final UserService userService;
 
-    public Birthday save(Birthday birthday);
+    public BirthdayService(BirthdayRepository birthdayRepository, UserService userService) {
+        this.birthdayRepository = birthdayRepository;
+        this.userService = userService;
+    }
+
+    public List<Birthday> getAllBirthdays() {
+
+        return birthdayRepository.findAll();
+    }
+
+    public List<Birthday> getBirthdaysByUserId(Long userId) {
+
+        return userService.getAllUsers().stream().filter(user -> user.getId().equals(userId)).findFirst().get().getBirthdays().stream().toList();
+    }
+
+    public Birthday save(Birthday birthday) {
+        return birthdayRepository.save(birthday);
+    }
 }
